backend/requirements.txt
fastapi
uvicorn[standard]
joblib
scikit-learn
pandas
numpy
PyPDF2
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
python-dotenv
openai
